# Benchstat results

```
goos: linux
goarch: amd64
pkg: github.com/hunterwilkins2/rsv
cpu: 13th Gen Intel(R) Core(TM) i7-13700K
            │   old.out   │               new.out               │
            │   sec/op    │   sec/op     vs base                │
Read-24       800.9n ± 3%   478.6n ± 3%  -40.24% (p=0.000 n=10)
CSVRead-24    681.8n ± 2%   674.2n ± 1%   -1.10% (p=0.002 n=10)
Write-24      673.2n ± 1%   513.5n ± 1%  -23.73% (p=0.000 n=10)
CSVWrite-24   622.7n ± 1%   635.6n ± 2%   +2.07% (p=0.000 n=10)
geomean       691.7n        569.7n       -17.64%

            │   old.out    │                new.out                 │
            │     B/op     │     B/op      vs base                  │
Read-24        1088.0 ± 0%     672.0 ± 0%  -38.24% (p=0.000 n=10)
CSVRead-24      664.0 ± 0%     664.0 ± 0%        ~ (p=1.000 n=10) ¹
Write-24      4.414Ki ± 0%   4.203Ki ± 0%   -4.78% (p=0.000 n=10)
CSVWrite-24   4.203Ki ± 0%   4.203Ki ± 0%        ~ (p=1.000 n=10) ¹
geomean       1.891Ki        1.656Ki       -12.43%
¹ all samples are equal

            │   old.out   │               new.out                │
            │  allocs/op  │ allocs/op   vs base                  │
Read-24        36.00 ± 0%   23.00 ± 0%  -36.11% (p=0.000 n=10)
CSVRead-24     16.00 ± 0%   16.00 ± 0%        ~ (p=1.000 n=10) ¹
Write-24      15.000 ± 0%   3.000 ± 0%  -80.00% (p=0.000 n=10)
CSVWrite-24    3.000 ± 0%   3.000 ± 0%        ~ (p=1.000 n=10) ¹
geomean        12.69        7.586       -40.21%
¹ all samples are equal
```
## Read Performance
`40.24%` increase in performance vs `v1.0.0`
`29.01%` increase in performance vs `csv.Reader`  

> Still need to figure out how to reduce allocations to match csv's #, but this is a start. csv 16 allocs/op. rsv 23 allocs/op

## Write Performance
`23.73%` increase in performance vs `v1.0.0`
`19.21%` increase in performance vs `csv.Writer`